linux 64 build updates

+++ src/variables.cpp
@@ -1,3 +1,5 @@
+#include <cmath>
+#include <numeric>
 #include <algorithm>
 
 #include <wx/datstrm.h>
@@ -909,7 +911,7 @@ wxString VarValue::AsString( wxChar arrsep, wxChar tabsep )
 	case VV_STRING: return m_str;
 	case VV_NUMBER:
 	{
-		if (std::isnan(m_val))
+		if (std::isnan((float)m_val))
 			return "NaN";
 		else
 			return wxString::Format("%g", (float)m_val);
